requirements.txt
PyQt5
psutil
pyqtgraph
python-dotenv
psycopg2